+++ Test/list/list_add_at.c
@@ -30,7 +30,7 @@ int main() {
         if (gListAddItemAt(list, &i, 0) == -1) {
             return st_errno;
         }
-        for (uint16_t j = 0; j <= i; ++j) {
+        for (unsigned int j = 0; j <= i; ++j) {
             printf("%d ", *(int *) gListGetItem(list, j));
         }
         printf("\n");
