+++ Test/list/list_remove.c
@@ -31,12 +31,12 @@ int main() {
         if (gListAddItem(list, &i) == -1) {
             return st_errno;
         }
-        for (uint16_t j = 0; j <= i; ++j) {
+        for (unsigned int j = 0; j <= i; ++j) {
             printf("%d ", *(int *) gListGetItem(list, j));
         }
         printf("\n");
     }
-    for (uint16_t i = 9; i >= 0; --i) {
+    for ( int i = 9; i >= 0; --i) {
 
         if (gListRemoveItem(list, i) == -1) {
             return st_errno;
